==> 2_Data Cleaning.ipynb <==
# Data Cleaning

## Importing Libraries

import pandas as pd
import numpy as np
from nltk.tokenize import RegexpTokenizer
from nltk.stem import WordNetLemmatizer

## Cleaning Retraction Data

### General Cleaning

retract = pd.read_csv('./total_retraction_data.csv', index_col=False)

As discussed in the "Data Collection" notebook, the querying script would have ideally run through the entire DOI column. However, I often had 404 errors or the script would stop abruptly with no error message. For this reason, I decided to break the DOI column into small sections to closely monitor the information and save the information I received frequently. Because of this, several CSV files were created, each labeled with what DOI index values were used when pulling the information. This information was then cleaned into one complete CSV file in a notebook that will not be provided. This CSV file is labeled as "total_retraction_data.csv" and was opened above. 

retract = retract.drop(columns=['Unnamed: 0', 'doi_check'])

retract = retract.dropna(axis=0, subset=['text'])

retract['retraction_binary'] = 1

retract.to_csv('./no_null_text_retraction_data.csv')
retract = pd.read_csv('./no_null_text_retraction_data.csv')
retract = retract.drop(columns=['Unnamed: 0'])

The dataframe was cleaned in several ways. Unnecessary columns were dropped. Any articles that had no text data were dropped. A new column named "retraction_binary" was created that was filled with the integer 1 to indicate that all of the articles in this dataframe were articles that had been retracted. Using .info method below, it is evident that the only remaining articles had full text. Once the data was cleaned, it was saved into a new CSV file for proofing.

retract.info()

### Unpacking Keywords

ls_keywords_list = []
count = 0
for i in retract['keywords']:
    keywords_list = []
    if i == []:
        ls_keywords_list.append([])
    else:
        for j in i.split():
            keywords_list.append(j.replace("'",'').replace('[','').replace(',','').replace(']','').replace('(','').replace(')','').replace('\\n', '').replace('\\n','').lower())
        if keywords_list == ['']:
            ls_keywords_list.append([])
        else:
            ls_keywords_list.append(keywords_list)

retract['unpacked_keywords'] = ls_keywords_list

The keywords list for each article was a string. In this format, this information is not useful. Because of this, the above script was used to "unpack" each keyword and place it into a true list. These lists were then placed into a new column, "unpacked_keywords".

### Cleaning and Lemmatizing Text

lemmatizer = WordNetLemmatizer()
tokenizer = RegexpTokenizer(r'\w+')

clean_text = []
clean_text_lem = []

for i in range(0, len(retract['text'])):
    ls_words = []
    ls_lem = []
    for j in tokenizer.tokenize(retract['text'][i]):
        try:
            int(j)
        except:
            if len(j) < 45:
                ls_words.append(j)
                ls_lem.append(lemmatizer.lemmatize(j))
            else:
                pass
        
    clean_text.append(' '.join(ls_words))
    clean_text_lem.append(' '.join(ls_lem))
    
    print(i)

retract['clean_text'] = clean_text
retract['clean_text_lem'] = clean_text_lem

The text of each article as it was originally received from the query needed to be cleaned in several ways. There were several symbols and numerical values that needed to be removed. To remove these characters, the RegexpTokenizer was used. Additionally, several "words" were in fact URLs or strings of characters that were not intelligible (as they may have been artifacts from mathematical equations in the article). Thus, any string that was above 45 characters was ignored. Once these cleaning steps had been taken, the text was saved in two different ways: one method with no further formatting, and one method where a lemmatizer was used. The now two different text bodies were then combined back together using spaces and added to new columns in the dataframe. Text that had no further formatting was saved as "clean_text" while lemmatized text was saved as "clean_text_lem".

## Cleaning No Retraction Data

### Correcting Indexing Error During PMC Querying

#reading all text CSV files from PMC querying
text_one = pd.read_csv('./text__no_retraction__0_5545.csv')
text_two = pd.read_csv('./text__no_retraction__5545_6166.csv')
text_three = pd.read_csv('./text__no_retraction__6166_end.csv')

#reading all abstract CSV files from PMC querying
abstract_one = pd.read_csv('./abstract__no_retraction__0_5545.csv')
abstract_two = pd.read_csv('./abstract__no_retraction__5545_6166.csv')
abstract_three = pd.read_csv('./abstract__no_retraction__6166_end.csv')

#reading all keyword CSV files from PMC querying
keywords_one = pd.read_csv('./keywords__no_retraction__0_5545.csv')
keywords_two = pd.read_csv('./keywords__no_retraction__5545_6166.csv')
keywords_three = pd.read_csv('./keywords__no_retraction__6166_end.csv')

#reading all publisher CSV files from PMC querying
publisher_one = pd.read_csv('./publisher__no_retraction__0_5545.csv')
publisher_two = pd.read_csv('./publisher__no_retraction__5545_6166.csv')
publisher_three = pd.read_csv('./publisher__no_retraction__6166_end.csv')

#reading all DOI CSV files from PMC querying
doi_one = pd.read_csv('./doi__no_retraction__0_5545.csv')
doi_two = pd.read_csv('./doi__no_retraction__5545_6166.csv')
doi_three = pd.read_csv('./doi__no_retraction__6166_end.csv')

As discussed in the "Data Collection" notebook, the script would have ideally run through the entire DOI column. However, when using this function, I often had 404 errors or the script would stop abruptly with no error message. For this reason, I decided to break the DOI column into three different sections to closely monitor the information and save the information I received more frequently. Because of this, three different CSV files were created, each labeled with what DOI index values were used when pulling the information. These CSV files are opened in the above script.

text_one.tail()

text_two.head()

text_two.tail()

text_three.head()

I realized I had duplicated two query calls when completing the PMC no retraction querying for the no retraction data. Because of this, I had to drop the appropriate rows from the CSV files for each piece of information determined while querying. The rows were dropped in the script below.

#dropping duplicate entries from text dataframes
text_one = text_one.drop(5545, axis=0)
text_two = text_two.drop(621, axis=0)

#dropping duplicate entries from abstract dataframes
abstract_one = abstract_one.drop(5545, axis=0)
abstract_two = abstract_two.drop(621, axis=0)

#dropping duplicate entries from keywords dataframes
keywords_one = keywords_one.drop(5545, axis=0)
keywords_two = keywords_two.drop(621, axis=0)

#dropping duplicate entries from publisher dataframes
publisher_one = publisher_one.drop(5545, axis=0)
publisher_two = publisher_two.drop(621, axis=0)

#combining all text dataframes
text_total = pd.concat([text_one, text_two, text_three], axis =0)

#combining all DOI dataframes
doi_total = pd.concat([doi_one, doi_two, doi_three], axis =0)

#combining all abstract dataframes
abstract_total = pd.concat([abstract_one, abstract_two, abstract_three], axis =0)

#combining all keyword dataframes
keywords_total = pd.concat([keywords_one, keywords_two, keywords_three], axis =0)

#combining all publisher dataframes
publisher_total = pd.concat([publisher_one, publisher_two, publisher_three], axis =0)

Each CSV file for each piece of information from the PMC query was concatenated together by row. Once the dataframe was created for each piece of information, the dataframe was then saved as a CSV file in the script below for proofing.

#saving all concatenated dataframes for proofing
text_total.to_csv('./plos_only_no_retraction_text.csv')
doi_total.to_csv('./plos_only_no_retraction_doi.csv')
abstract_total.to_csv('./plos_only_no_retraction_abstract.csv')
keywords_total.to_csv('./plos_only_no_retraction_keywords.csv')
publisher_total.to_csv('./plos_only_no_retraction_publisher.csv')

#rereading all concatenated dataframes to continue working
text = pd.read_csv('./plos_only_no_retraction_text.csv')
abstract = pd.read_csv('./plos_only_no_retraction_abstract.csv')
keywords = pd.read_csv('./plos_only_no_retraction_keywords.csv')
publisher = pd.read_csv('./plos_only_no_retraction_publisher.csv')
doi = pd.read_csv('./plos_only_no_retraction_doi.csv')

### Combining Data from PMC Queries

#pulling all text data for no retractions
ls_text = []
for i in text['0']:
    ls_text.append(i)
print(len(ls_text))

#pulling all abstract data for no retractions
ls_abstract = []
for i in abstract['0']:
    ls_abstract.append(i)
print(len(ls_abstract))

#pulling all keywords list data for no retractions
ls_keywords = []
for i in keywords['0']:
    ls_keywords.append(i)
print(len(ls_keywords))

#pulling all publisher data for no retractions
ls_publisher = []
for i in publisher['0']:
    ls_publisher.append(i)
print(len(ls_publisher))

#pulling all DOI data for no retractions
ls_doi = []
for i in doi['1']:
    ls_doi.append(i)
print(len(ls_doi))

To place all of the information from the PMC no retraction query, the information from the newly formed CSV files was pulled and placed into a list for each information piece. The PubMed no retraction CSV file was then opened. All of the lists were concatenated with the PubMed no retraction dataframe to combine all of the no retraction inforamtion into a single dataframe.

no_retract = pd.read_csv('./pubmed_data_second_no_retraction.csv')

no_retract = pd.concat([no_retract, pd.Series(ls_text), pd.Series(ls_abstract), pd.Series(ls_keywords), 
                        pd.Series(ls_publisher), pd.Series(ls_doi)], axis=1)

no_retract = no_retract.drop(columns=['Unnamed: 0'])

no_retract = no_retract.rename(columns={'0':'id', '1':'doi', '2':'language', '3':'year', '4':'month', '5':'day', 
                      '6':'volume', '7':'issue', '8':'journal', '9':'title', '10':'page', 
                      0:'text', 1:'abstract', 2:'keywords', 3:'publisher', 4:'doi_check'})

Unnecessary columns were dropped. For the columns that remain, the columns were renamed to the appropriate information based on the order in which the information was pulled during querying. The complete dataframe can be seen below.

no_retract['doi'][5540:5550]

no_retract['doi_check'][5540:5550]

no_retract.to_csv('./plos_only_no_retraction_data.csv')

The DOI values were compared between the two different DOI columns to ensure that the concatenating between the different dataframes was completed accurately. Once it was determined that there were no errors in concatenating, the dataframe was saved as a new CSV file for proofing.

### General Cleaning

no_retract = pd.read_csv('./plos_only_no_retraction_data.csv', index_col=False)
no_retract = no_retract.drop(columns=['Unnamed: 0', 'doi_check'])

print(no_retract['abstract'].isnull().sum())
print(no_retract['text'].isnull().sum())

The above script shows the number of null values in the "abstract" and "text" column. In the case that there is simply too much text in the full article to be used for NLP modeling, it may be possible to use the information in the abstract of each article.

no_retract = no_retract.dropna(axis=0, subset=['text'])

Any articles that were missing text were dropped from the dataframe.

np.random.choice(no_retract['doi'], 50)

50 DOIs of articles that were not retracted from PLOS ONE were pulled. These DOIs were used to manually check that the articles pulled during the query were in fact not retracted. While all of the articles proved to be not retracted, 5 of the articles had actually been corrected. The reasons for correction were as followed: author byline changes and fixing a citation, supporting figure appears incorrectly, authors spelled incorrectly, errors in figures, and affiliation listing error. Corrections are not the same as retractions, but may be something to look into in the future.

no_retract['retraction_binary'] = 0
print(no_retract['retraction_binary'].value_counts())
no_retract.head()

no_retract.to_csv('./no_null_text_plos_only_no_retraction_data.csv')
no_retract = pd.read_csv('./no_null_text_plos_only_no_retraction_data.csv')
no_retract = no_retract.drop(columns=['Unnamed: 0'])

A new column named "retraction_binary" was created that was filled with the integer 0 to indicate that all of the articles in this dataframe were articles that had not been retracted. Using .info method below, it is evident that the only remaining articles had full text. Once the data was cleaned, it was saved into a new CSV file for proofing.

no_retract.info()

### Unpacking Keywords

no_retract['keywords'].value_counts()

I went to unpack the keywords list as I had done for the retraction data. However, after completing .value_counts method for the column, I realized that only one data point out of 9214 articles actually had a keywords list. I manually checked using the same 50 random DOI values to determine that none of the articles randomly chosen actually had a keywords list. Because of this, I did not bother to unpack the keywords list.

### Cleaning and Lemmatizing Text

lemmatizer = WordNetLemmatizer()
tokenizer = RegexpTokenizer(r'\w+')

clean_text = []
clean_text_lem = []

for i in range(0, len(no_retract['text'])):
    ls_words = []
    ls_lem = []
    for j in tokenizer.tokenize(no_retract['text'][i]):
        try:
            int(j)
        except:
            if len(j) < 45:
                ls_words.append(j)
                ls_lem.append(lemmatizer.lemmatize(j))
            else:
                pass
        
    clean_text.append(' '.join(ls_words))
    clean_text_lem.append(' '.join(ls_lem))
    
    print(i)

print(len(no_retract['text'][1]))
print(len(clean_text[1]))
print(len(clean_text_lem[1]))

no_retract['clean_text'] = clean_text
no_retract['clean_text_lem'] = clean_text_lem
no_retract.head()

The text of each article as it was originally received from the query needed to be cleaned in several ways. There were several symbols and numerical values that needed to be removed. To remove these characters, the RegexpTokenizer was used. Additionally, several "words" were in fact URLs or strings of characters that were not intelligible (as they may have been artifacts from mathematical equations in the article). Thus, any string that was above 45 characters was ignored. Once these cleaning steps had been taken, the text was saved in two different ways: one method with no further formatting, and one method where a lemmatizer was used. The now two different text bodies were then combined back together using spaces and added to new columns in the dataframe. Text that had no further formatting was saved as "clean_text" while lemmatized text was saved as "clean_text_lem".

no_retract.loc[no_retract['year']==2014].index

no_retract = no_retract.drop(range(43, 167))
no_retract['year'].value_counts()

During the querying process, the dates randomly chosen were used as the end point of the query. Thus, 167 articles were pulled where each date is the last possible date the article could have been published on. Because of this, there were 124 articles that were published in 2014. For consistency, these articles were dropped.

## Cleaning Combined Datasets

total = pd.concat([retract, no_retract], axis=0)
total

The retraction and no retraction dataframes were concatenated together to create a dataframe that has all of the inforamtion from all of the queries completed in the "Data Collection" notebook.

total = total.drop(columns=['id', 'language', 'publisher'])

total = total.drop(columns=['page'])

total['year'] = total['year'].astype(str)
total['month'] = total['month'].astype(str)
total['day'] = total['day'].astype(str)
total['volume'] = total['volume'].astype(str)
total['issue'] = total['issue'].astype(str)

Unnecessary columns were dropped. The "year," "month," "day," "volume," and "issue" columns were retyped as strings so that they could be dummied in later notebooks. Once cleaned, the dataframe was saved into a new CSV file for proofing.

total.to_csv('./total_plos_only_data.csv')
total = pd.read_csv('./total_plos_only_data.csv')

total['doi'].value_counts().head(25)

total = total.drop_duplicates(subset='doi', keep='first')

During the querying process, some DOI values were repeated even though duplicates should have been dropped before pulling from the PMC database. These DOI values are unique strings that identify a specific article. Because of this, if a DOI is repeated, then the article text is repeated within the dataframe as well. Thus, these duplicates were dropped and saved to a new CSV file for proofing.

total.to_csv('./total_plos_only_data_no_duplicates.csv')
total = pd.read_csv('./total_plos_only_data_no_duplicates.csv')

total['clean_text'][950:955]

During the tokenizing process, one text must have been removed completely (as the text may not have had any spaces). This row was dropped as it serves no purpose to modeling.

total.to_csv('./total_data_plos_only_cleaned.csv')
total[1537:].to_csv('./no_retraction_data_plos_only_cleaned.csv')
total[:1537].to_csv('./retraction_data_plos_only_cleaned.csv')

Because of the way the dataframes were concatenated, the dataframe from index 0-1537 are only retracted articles while the index 1537-10643 are only non-retracted articles. Thus, the final dataframe was saved in three different ways: the complete dataframe, only retracted articles, and only non-retracted articles.
